requirements.txt
amqp==5.3.1
asgiref==3.11.0
billiard==4.2.4
celery==5.3.6
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
coverage==7.13.0
Django==5.1
django-cors-headers==4.3.1
django-filter==23.5
django-redis==5.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
Faker==22.6.0
gunicorn==21.2.0
inflection==0.5.1
iniconfig==2.3.0
kombu==5.6.1
mysql-connector-python==9.5.0
mysqlclient==2.2.7
packaging==25.0
pillow==12.0.0
pluggy==1.6.0
prompt_toolkit==3.0.52
PyJWT==2.10.1
PyMySQL==1.1.2
pytest==7.4.4
pytest-cov==7.0.0
pytest-django==4.7.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.3
redis==5.0.1
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.4
tzdata==2025.2
uritemplate==4.2.0
vine==5.1.0
wcwidth==0.2.14
wheel==0.45.1
